Backend/requirements.txt
pandas
numpy 
fastapi 
uvicorn 
requests 
langchain 
openai
psycopg2-binary
mysql-connector-python 
sqlalchemy
cryptography
chromadb
python-dotenv
